refactor: extract outlier detection hooks/helpers
