+++ src/pages/OutlierDetection/OutlierDetection.js
@@ -7,128 +7,14 @@ import MaxResultsAlertBar from '../../components/MaxResultsAlertBar/MaxResultsAl
 import PageHeader from '../../components/PageHeader/PageHeader'
 import { useSidebar } from '../../components/Sidebar/SidebarContext'
 import { convertDateToApiDateFormat } from '../../helpers/dates'
-import threeMonthsAgo from '../../helpers/threeMonthsAgo'
 import { apiConf } from '../../server.conf'
 import cssPageStyles from '../Page.module.css'
-import {
-    Z_SCORE,
-    DEFAULT_THRESHOLD,
-    DEFAULT_ALGORITHM,
-    DEFAULT_MAX_RESULTS,
-    DEFAULT_ORDER_BY,
-} from './constants'
+import { Z_SCORE } from './constants'
+import convertElementFromApiResponse from './convert-element-from-api-response'
+import convertElementToToggleFollowupRequest from './convert-element-to-toggle-followup-request'
 import Form from './Form'
 import OutlierAnalyisTable from './OutlierAnalysisTable/OutlierAnalysisTable'
-
-const getElementDisplayName = e => {
-    let str = e.deName
-
-    // In the context of a dataElement, the default COC or AOC means "none".
-    // The "default" string is not localised, and probably won't ever be.
-    // That is why the conditions below should work in the foreseeable future.
-    if (e.cocName !== 'default') {
-        str += ` (${e.cocName})`
-    }
-
-    if (e.aocName !== 'default') {
-        str += ` (${e.aocName})`
-    }
-
-    return str
-}
-
-const convertElementFromApiResponse = e => ({
-    displayName: getElementDisplayName(e),
-    key: `${e.aoc}-${e.coc}-${e.de}-${e.pe}-${e.ou}`,
-    marked: e.followup,
-    ...e,
-})
-
-const convertElementToToggleFollowupRequest = e => ({
-    dataElement: e.de,
-    period: e.pe,
-    orgUnit: e.ou,
-    categoryOptionCombo: e.coc || null,
-    attributeOptionCombo: e.aoc || null,
-    followup: !e.marked,
-})
-
-const useFormState = () => {
-    const [startDate, setStartDate] = useState(threeMonthsAgo())
-    const [endDate, setEndDate] = useState(new Date())
-    const [organisationUnitIds, setOrganisationUnitIds] = useState([])
-    const [algorithm, setAlgorithm] = useState(DEFAULT_ALGORITHM)
-    const [showAdvancedZScoreFields, setShowAdvancedZScoreFields] = useState(
-        false
-    )
-    const [threshold, setThreshold] = useState(DEFAULT_THRESHOLD)
-    const [orderBy, setOrderBy] = useState(DEFAULT_ORDER_BY)
-    const [dataStartDate, setDataStartDate] = useState(null)
-    const [dataEndDate, setDataEndDate] = useState(null)
-    const [dataSetIds, setDataSetIds] = useState([])
-    const [maxResults, setMaxResults] = useState(DEFAULT_MAX_RESULTS)
-
-    const handleStartDateChange = (event, date) => {
-        setStartDate(new Date(date))
-    }
-    const handleEndDateChange = (event, date) => {
-        setEndDate(new Date(date))
-    }
-    const handleAlgorithmChange = (event, index, value) => {
-        setAlgorithm(value)
-    }
-    const handleToggleAdvancedZScoreFields = () => {
-        if (showAdvancedZScoreFields) {
-            setDataStartDate(null)
-            setDataEndDate(null)
-            setOrderBy(DEFAULT_ORDER_BY)
-        }
-        setShowAdvancedZScoreFields(!showAdvancedZScoreFields)
-    }
-    const handleThresholdChange = (event, index, value) => {
-        setThreshold(value)
-    }
-    const handleOrderByChange = (event, index, value) => {
-        setOrderBy(value)
-    }
-    const handleDataStartDateChange = date => {
-        setDataStartDate(date && new Date(date))
-    }
-    const handleDataEndDateChange = date => {
-        setDataEndDate(date && new Date(date))
-    }
-    const handleDataSetsChange = ({ selected }) => {
-        setDataSetIds(selected)
-    }
-    const handleMaxResultsChange = (event, index, value) => {
-        setMaxResults(value)
-    }
-
-    return {
-        organisationUnitIds,
-        handleOrganisationUnitChange: setOrganisationUnitIds,
-        startDate,
-        handleStartDateChange,
-        endDate,
-        handleEndDateChange,
-        algorithm,
-        handleAlgorithmChange,
-        showAdvancedZScoreFields,
-        handleToggleAdvancedZScoreFields,
-        threshold,
-        handleThresholdChange,
-        orderBy,
-        handleOrderByChange,
-        dataStartDate,
-        handleDataStartDateChange,
-        dataEndDate,
-        handleDataEndDateChange,
-        dataSetIds,
-        handleDataSetsChange,
-        maxResults,
-        handleMaxResultsChange,
-    }
-}
+import useFormState from './use-form-state'
 
 const OutlierDetection = () => {
     const sidebar = useSidebar()
