+++ src/pages/OutlierDetection/convert-element-to-toggle-followup-request.js
@@ -1,0 +1,10 @@
+const convertElementToToggleFollowupRequest = e => ({
+    dataElement: e.de,
+    period: e.pe,
+    orgUnit: e.ou,
+    categoryOptionCombo: e.coc || null,
+    attributeOptionCombo: e.aoc || null,
+    followup: !e.marked,
+})
+
+export default convertElementToToggleFollowupRequest
